Move subcommand tests
