+++ clikt/src/commonTest/kotlin/com/github/ajalt/clikt/core/CliktCommandTest.kt
@@ -232,94 +232,6 @@ class CliktCommandTest {
     }
 
     @Test
-    @JsName("subcommand_cycle")
-    fun `subcommand cycle`() {
-        val root = TestCommand(called = false)
-        val a = TestCommand(called = false, name = "a")
-        val b = TestCommand(called = false)
-
-        shouldThrow<IllegalStateException> {
-            root.subcommands(a.subcommands(b.subcommands(a))).parse("a b a")
-        }.message shouldBe "Command a already registered"
-    }
-
-    @Test
-    @JsName("multiple_subcommands")
-    fun `multiple subcommands`() = forall(
-            row("foo a", 1, 0, "a", null, null),
-            row("foo a foo b", 2, 0, "b", null, null),
-            row("bar a", 0, 1, null, null, "a"),
-            row("bar a bar b", 0, 2, null, null, "b"),
-            row("bar --opt=o a", 0, 1, null, "o", "a"),
-            row("foo a bar --opt=o b foo c bar d", 2, 2, "c", null, "d"),
-            row("foo a bar b foo c bar --opt=o d", 2, 2, "c", "o", "d")
-    ) { argv, fc, bc, fa, bo, ba ->
-        val foo = MultiSub1(count = fc)
-        val bar = MultiSub2(count = bc)
-        val c = TestCommand(allowMultipleSubcommands = true).subcommands(foo, bar, TestCommand(called = false))
-        c.parse(argv)
-        if (fc > 0) foo.arg shouldBe fa
-        bar.opt shouldBe bo
-        if (bc > 0) bar.arg shouldBe ba
-    }
-
-    @Test
-    @JsName("multiple_subcommands_with_nesting")
-    fun `multiple subcommands with nesting`() {
-        val foo = MultiSub1(count = 2)
-        val bar = MultiSub2(count = 2)
-        val c = TestCommand(allowMultipleSubcommands = true).subcommands(foo.subcommands(bar))
-        c.parse("foo f1 bar --opt=1 b1 foo f2 bar b2")
-        foo.arg shouldBe "f2"
-        bar.opt shouldBe null
-        bar.arg shouldBe "b2"
-    }
-
-    @Test
-    @JsName("multiple_subcommands_nesting_the_same_name")
-    fun `multiple subcommands nesting the same name`() {
-        val bar1 = MultiSub2(count = 2)
-        val bar2 = MultiSub2(count = 2)
-        val c = TestCommand(allowMultipleSubcommands = true).subcommands(bar1.subcommands(bar2))
-        c.parse("bar a11 bar a12 bar a12 bar --opt=o a22")
-        bar1.arg shouldBe "a12"
-        bar2.opt shouldBe "o"
-        bar2.arg shouldBe "a22"
-    }
-
-    @Test
-    @JsName("multiple_subcommands_with_varargs")
-    fun `multiple subcommands with varargs`() = forall(
-            row("foo f1 baz", 1, 1, "f1", emptyList()),
-            row("foo f1 foo f2 baz", 2, 1, "f2", emptyList()),
-            row("baz foo", 0, 1, "", listOf("foo")),
-            row("baz foo baz foo", 0, 1, "", listOf("foo", "baz", "foo")),
-            row("foo f1 baz foo f2", 1, 1, "f1", listOf("foo", "f2"))
-    ) { argv, fc, bc, fa, ba ->
-        class Baz : TestCommand(name = "baz", count = bc) {
-            val arg by argument().multiple()
-        }
-
-        val foo = MultiSub1(count = fc)
-        val baz = Baz()
-        val c = TestCommand(allowMultipleSubcommands = true).subcommands(foo, baz)
-        c.parse(argv)
-
-        if (fc > 0) foo.arg shouldBe fa
-        baz.arg shouldBe ba
-    }
-
-    @Test
-    @JsName("multiple_subcommands_nesting_multiple_subcommands")
-    fun `multiple subcommands nesting multiple subcommands`() {
-        val c = TestCommand(allowMultipleSubcommands = true)
-                .subcommands(TestCommand(allowMultipleSubcommands = true))
-        shouldThrow<IllegalArgumentException> {
-            c.parse("")
-        }.message shouldContain "allowMultipleSubcommands"
-    }
-
-    @Test
     @JsName("treat_unknown_options_as_arguments")
     fun `treat unknown options as arguments`() {
         class C : TestCommand(treatUnknownOptionsAsArgs = true) {
@@ -351,13 +263,4 @@ class CliktCommandTest {
             C(false).parse("-fgi")
         }.message shouldBe "no such option: \"-g\"."
     }
-}
-
-private class MultiSub1(count: Int) : TestCommand(name = "foo", count = count) {
-    val arg by argument()
-}
-
-private class MultiSub2(count: Int) : TestCommand(name = "bar", count = count) {
-    val opt by option()
-    val arg by argument()
 }
